Rename from bbn to hume inside module
