+++ indra/sources/hume/__init__.py
@@ -1,1 +1,1 @@
-from .bbn_api import process_json_file_old, process_jsonld_file
+from .api import process_json_file_old, process_jsonld_file
